+++ src/scout_apm/instruments/jinja2.py
@@ -34,7 +34,6 @@ have_patched_template_render_async = False
 
 
 def ensure_installed():
-    global have_patched_environment_init
     global have_patched_template_render
 
     logger.debug("Instrumenting Jinja2.")
@@ -43,17 +42,7 @@ def ensure_installed():
         logger.debug("Couldn't import jinja2.Template - probably not installed.")
         return
 
-    if not have_patched_environment_init:
-        try:
-            Environment.__init__ = wrapped_environment_init(Environment.__init__)
-        except Exception as exc:
-            logger.warning(
-                "Failed to instrument jinja2.Environment.__init__: %r",
-                exc,
-                exc_info=exc,
-            )
-        else:
-            have_patched_environment_init = True
+    instrument_render_async()
 
     if not have_patched_template_render:
         try:
@@ -66,6 +55,39 @@ def ensure_installed():
             have_patched_template_render = True
 
 
+def instrument_render_async():
+    global have_patched_environment_init
+    global have_patched_template_render_async
+
+    if wrapped_render_async is None:
+        return
+
+    if not have_patched_environment_init and not hasattr(Template, "render_async"):
+        try:
+            Environment.__init__ = wrapped_environment_init_jinja_v2(
+                Environment.__init__
+            )
+        except Exception as exc:
+            logger.warning(
+                "Failed to instrument jinja2.Environment.__init__: %r",
+                exc,
+                exc_info=exc,
+            )
+        else:
+            have_patched_environment_init = True
+    elif hasattr(Template, "render_async") and not have_patched_template_render_async:
+        try:
+            Template.render_async = wrapped_render_async(Template.render_async)
+        except Exception as exc:
+            logger.warning(
+                "Failed to instrument jinja2.Template.render_async: %r",
+                exc,
+                exc_info=exc,
+            )
+        else:
+            have_patched_template_render_async = True
+
+
 @wrapt.decorator
 def wrapped_render(wrapped, instance, args, kwargs):
     tracked_request = TrackedRequest.instance()
@@ -75,12 +97,14 @@ def wrapped_render(wrapped, instance, args, kwargs):
 
 
 @wrapt.decorator
-def wrapped_environment_init(wrapped, instance, args, kwargs):
+def wrapped_environment_init_jinja_v2(wrapped, instance, args, kwargs):
     """
     Delayed wrapping of render_async(), since Template won't have this method
     until after jinja2.asyncsupport is imported, which since Jinja2 2.11.0 is
     done conditionally in Environment.__init__:
     https://github.com/pallets/jinja/issues/765
+
+    This is no longer needed since Jinja2 v3.0.0
     """
     global have_patched_template_render_async
     result = wrapped(*args, **kwargs)