+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Pending
 
 ### Added
+- Added `terms_enum` instrumentation for ElasticSearch 7.14.0.
+  ([Issue #671](https://github.com/scoutapp/scout_apm_python/issues/671))
 
 ### Fixed